Separate hurl tests for authenticated and non-authenticated tests
With the addition of JWT authentication to SMD and BSS (and the option to
enable/disable it), there needs to be a way to test both.

In this commit, the original ochami-services.yml (where the services do not use
JWT auth) is moved to ochami-services-noauth.yml and ochami-services.yml is
used for running the services with JWT auth. The hurl tests are separated into
'auth' and 'noauth' directories. For ochami-services.yml to properly use
authentication, hydra.yml must also be running.

The 'auth' tests test against ochami-services.yml, but these only test that not
presenting a token to protected endpoints yields a 401 (Unauthorized) response.
Therefore, nothing is modified.

The 'noauth' tests test against ochami-services.yml (which runs against the
postgres DB in ochami-services.yml) and perform all of the SMD and BSS endpoint
testing. These tests actually modify the postgres DB.